requirements.txt
ExifRead~=3.0.0
fastapi~=0.111.1
numpy~=1.26.4
pgvector~=0.3.2
pillow~=10.4.0
psycopg2-binary~=2.9.9
pydantic~=2.8.2
python-dotenv~=1.0.1
sqlalchemy~=2.0.31
sqlmodel~=0.0.21
starlette~=0.38.1
sentence-transformers~=3.0.1
uvicorn~=0.30.3
